completed s3 bucket upload with return image url
