+++ package.json
@@ -12,8 +12,7 @@
     "author": "",
     "license": "ISC",
     "dependencies": {
-        "@aws-sdk/client-s3": "^3.501.0",
-        "aws-sdk": "^2.1547.0",
+        "@aws-sdk/client-s3": "^3.503.1",
         "bcryptjs": "^2.4.3",
         "cookie-parser": "^1.4.6",
         "cors": "^2.8.5",
